fix: circleci doesn't have permissions (#34)
CircleCI runs a process called '.machine-agent' which needs permissions to do all the things required for the test.

I created a variable to make it clearer what this path is for.
The macos orb[1] has an 'add-permission' command but it's hard coded to expect a bundle id, not a path to an executable so we can't use it.

Some permissions can be got from

      - macos/add-uitest-permissions
      - macos/add-safari-permissions
      - macos/add-voiceover-permissions

but not enough for a VoiceOver test

[1]: https://circleci.com/developer/orbs/orb/circleci/macos